+++ stage0/.cargo/config.toml
@@ -2,7 +2,7 @@
 target = "x86_64-unknown-none"
 
 [target.x86_64-unknown-none]
-rustflags = "-C relocation-model=static -C code-model=large -C linker-plugin-lto"
+rustflags = "-C relocation-model=static -C code-model=large"
 
 [unstable]
 # We need to build std / core because of the code model. This may be removed if / when it's possible to build stage0 without it in the future.
