+++ stage0/src/main.rs
@@ -20,13 +20,7 @@
 
 extern crate alloc;
 
-use alloc::vec;
-use core::{
-    arch::asm,
-    ffi::{c_void, CStr},
-    mem::MaybeUninit,
-    panic::PanicInfo,
-};
+use core::{arch::asm, ffi::c_void, mem::MaybeUninit, panic::PanicInfo};
 use goblin::elf::header;
 use oak_sev_guest::io::PortFactoryWrapper;
 use static_alloc::bump::Bump;
@@ -69,9 +63,6 @@ static SEV_CPUID: MaybeUninit<oak_sev_guest::cpuid::CpuidPage> = MaybeUninit::un
 
 /// We create an identity map for the first 1GiB of memory.
 const TOP_OF_VIRTUAL_MEMORY: u64 = Size1GiB::SIZE;
-
-/// The file path used by Stage0 to read the kernel command-line from the fw_cfg device.
-const CMDLINE_FILE_PATH: &[u8] = b"opt/stage0/cmdline\0";
 
 extern "C" {
     #[link_name = "pd_addr"]
@@ -294,24 +285,7 @@ pub extern "C" fn rust64_start(encrypted: u64) -> ! {
         zero_page.add_setup_data(setup_data);
     }
 
-    let cmdline_path = CStr::from_bytes_with_nul(CMDLINE_FILE_PATH).expect("invalid c-string");
-    if let Some(cmdline_file) = fwcfg.find(cmdline_path) {
-        let cmdline_size = cmdline_file.size();
-        // Make the buffer one byte longer so that the kernel command-line is null-terminated.
-        let buf = vec![0u8; cmdline_size + 1].leak();
-        let actual_size = fwcfg
-            .read_file(&cmdline_file, buf)
-            .expect("could not read cmdline");
-        assert_eq!(
-            actual_size, cmdline_size,
-            "cmdline size did not match expected size"
-        );
-
-        let cmdline = CStr::from_bytes_with_nul(buf).expect("invalid kernel command-line");
-        log::debug!(
-            "Kernel cmdline: {}",
-            cmdline.to_str().expect("invalid kernel commande-line")
-        );
+    if let Some(cmdline) = kernel::try_load_cmdline(&mut fwcfg) {
         zero_page.set_cmdline(cmdline);
     }
